examples: Turn 'Dialer Example' into 'Handy Widget Factory'
Refactor 'Dialer Example' so it has a proper window class with a
GtkBuilder descriptor file, and turn it into 'Handy Widget Factory'
which will allow to present more widgets from the library in one
convenient demo application.

The new application adds a sidebar to switch between the demoed widgets
and a welcome page.