tests: Use gtk_test_init()
Replace calls to g_test_init() and gtk_init() by calls to
gtk_test_init().

This better reflects the goal (testing GTK+ code), this makes the code
simpler and it does a bit more to make the tests deterministic like
setting the locale to C. This also allows us to automatically benefit
from improvements and fixes in the GTK+ test initialization process.