dialer: Make the grid visible and forbid show all
Without making the grid visible, usage of show_all() is required for the
dialer to be visible, but usage of show_all() breaks control over the
visibility of the action buttons so its usage should be forbidden.

This fixes both the problems by fixing the visibility rules of the grid.