leaflet: Init mode transition positions to the final values
Set the target and current mode transition positions to the final 1.0
value instead of the initial 0.0 one.

This fixes an inconsistency on initially unfolded leaflets as they would
be drawn in their folded state with no animation triggered to transition
to the final state as the target and the current values were the same.