example: Use separators to separate the panels
Use a GtkSeparator to separate the stack sidebar from the various demo
widgets, use custom styling to make the sidebar not draw its own
separator, and move the start of the header bar and the header bar
separator out of their box and directly into the leaflet.

This allows to have the separators well aligned with no workaround other
than a simple custom style, to have no separator appearing when the
window is folded, and to not trigger the GTK+ bug making titlebar
separators appear black.