header-group: Ref itself instead of header bars
When adding a header bar, ref the header group and connect to 'destroy'
signal of the header bar. When a header bar is destroyed or
hdy_header_group_remove_header_bar() is called, unref the header bar and
remove it from the list.

This way, a non-empty header group is only destroyed after every header
bar it contains has been removed from the group or destroyed.

Fixes #56