leaflet: Fix the folding sliding children padding
Sets the children padding of the folding sliding animation depending on
the surface they'll be drawn on.

This doesn't change a thing for the sliding animation, but this will
avoid the children to be moved when snapshotting them, which is needed
for the over and under animations — which will be added in the next
commit — to work correctly.