preferences-window: Hide filtered-out rows
The CSS engine works in such a way that invisible children are treated
as visible widgets, which breaks the expectations of the .preferences
style class when filtering a row, leading to straight corners when the
first row or last row are filtered out.

This works around it by hiding filtered out widgets.