build: Add an override for pygobject
PyGObject seems to load libraries lazily, which means the constructor won't
run until after we've referenced a libhandy symbol. However, this won't
work if a type is first referenced from a GtkBuilder file, so we need to
make sure the constructor runs before that. Add an override to preload it.

Thanks to Zander Brown for explanations.

Signed-off-by: Alexander Mikhaylenko <alexm@gnome.org>