keypad: Fix the code style
Use a single character per line in function definition prototypes, use
newlines to separate preconditions and the retrieval of the private
fields from the rest, put spaces before opening parentheses, and drop
lines empty lines in switches after a final break.

This makes the code style more in line with the rest of the code base.