swipe-tracker: Move some event handling to bubble stage
The reason for using capture phase at first was because we don't get stop
events on the bubble phase. However, with capture phase we also invert the
nested widget handling.

Instead, try to still get scroll stop events on the capture phase, while
handling everything else on the bubble phase.

Signed-off-by: Alexander Mikhaylenko <alexm@gnome.org>