stackable-box: Fix allocation during mode transitions
On the very last frame the size doesn't change anymore, and
queue_resize() fails. Use get_allocate() instead, also clean up the
unused comments.

This will help to prevent glitches later with HdyHeaderGroup in
combination with
https://gitlab.gnome.org/GNOME/libhandy/-/merge_requests/480

Since all children will be visible during mode transitions,
HdyHeaderGroup will switch the decoration layout on the last frame of
the transition, and it's important that it gets a relayout it needs.

Signed-off-by: Alexander Mikhaylenko <alexm@gnome.org>