shadow-helper: Use OVER operator instead of ATOP
Since GTK draws cairo content in an offscreen surface when it's to be
overlaid onto OpenGL content, and that surface has alpha channel is empty
at first, using ATOP here results in nothing being drawn. So here we have
to finally stop doing that and use plain OVER. While it results in a subtle
glitch with window corners in case of GtkWindow, we have a solution for
that now, and there's not much reason to keep using it.

Fixes https://gitlab.gnome.org/GNOME/libhandy/issues/242

Signed-off-by: Alexander Mikhaylenko <alexm@gnome.org>