swipeable: Add params to get_swipe_area()
Add a navigation direction param and a gesture type param to
get_swipe_area(). This allows having different swipe areas for swipes
back and forward, and for drag swipes and non-drag swipes, which is
needed to support e.g. swiping back from anywhere and swiping forward
from the border with an 'over' transition on touchscreens, while
changing nothing on touchpads.