Add a separate header for HdySwipeTracker forward declaration
HdySwipeable and HdySwipeTracker reference each other, so one of them needs
a forward declaration. Previously it was in hdy-swipeable.h, but it seems
to confuse gtk-doc.

To work around that, mirror what GTK does: have a separate public header
named hdy-types.h for forward declarations, include it in hdy-swipeable.h,
and mark it as private header for gtk-doc.

Fixes https://gitlab.gnome.org/GNOME/libhandy/-/issues/302

Signed-off-by: Alexander Mikhaylenko <alexm@gnome.org>