stackable-box: Don't use last visible child's size when it's NULL
When interpolate-size is true, we interpolate between the sizes of the
current visible child and last visible child. This works assuming that the
last visible child exists. However, before we switch visible child the
first time, it doesn't, so we get 0 size and interpolate between that and
the actual size.

Instead, use regular visible child size if last visible child is null.

Signed-off-by: Alexander Mikhaylenko <alexm@gnome.org>