swipe-tracker: Fix coordinate transformation for scrolling
Event coordinates are relative to the GdkWindow the event was on.
Meanwhile, gtk_widget_translate_coordinates() transforms coordinates from
a widget to another widget, instead of from a window to widget. This means
that it's only correct if the widget and window coordinates match, which is
not always the case and leads to misplaced swipe area in some cases.