Add AdwBin
GTK4 removed GtkBin as a part of the GtkContainer removal, with the
argument being that bin is easy to reimplement anyway. However, while it's
easy enough, it's still quite a lot of boilerplate that needs to be copied
around, and involves overriding dispose() which is generally not something
you're expected to do at all when using bindings.

Unlike GtkBin, AdwBin is also not abstract and can be used as is, replacing
the cases where in GTK3 you'd use GtkEventBox as a child placeholder.

Fixes https://gitlab.gnome.org/exalm/libadwaita/-/issues/32