preferences-window: Bind search model lazily
The row state depends on the number of pages in the window, and we weren't
updating the rows if it changes.

While we can notify all rows when it happens, a much easier and more
efficient way is to instead bind the model lazily when we start search.
Additionally, we can unbind it when it ends so we don't keep the rows
around when not needed.