Import Adwaita stylesheet
Install it into /org/gtk/libgtk/theme so that GTK can find it. Make
HighContrast and HighContrastInverse into :hc and :hc-dark variants, alias
each variant as Adwaita-hc etc themes.

Ensure Adwaita is loaded instead of the default, take care to load
Adwaita-hc when HighContrast is selected system-wide.

Use Adwaita-hc-dark as a dark variant of the high contrast mode.

Don't integrate the previous libadwaita styles into the stylesheet yet,
that can be done separately and this is a big change as is.

Fixes https://gitlab.gnome.org/GNOME/libadwaita/-/issues/91