avatar: Replace image load func with paintables
Introduce #AdwAvatar:custom-image. The original motivation for having image
load functions was to make it work with HiDPI which paintables already can
do.

Since then, libhandy has introduced a GLoadableIcon-based API to allow
async loading, but paintables are generic enough that it should be possible
to implement it as well.