stylesheet: Transparentize header bar contents for backdrop
The goal with dropping backdrop styles for GTK4 was to restrict backdrop
for header bars. However, it's quite complicated and also inconsistent, for
example leaving window titles unchanged, or removing hover effect from
regular buttons, but leaving on flat buttons.

Instead, go with a simpler approach, making header bar contents 50%
transparent. This produces a visually similar result to what it was before,
but is much simpler.

Unlike the previous method with separate styles, this leaves the hover
styles unchanged, but still toned down with the rest of the header bar.

Fixes https://gitlab.gnome.org/GNOME/libadwaita/-/issues/107