stylesheet: Simplify checks and radios
Use flatter checks, without borders. Use distinct enough background to make
them visible in dark.

Unify the appearance of regular and selection mode checks, at this point
selection mode is not much more than a larger variant.

Keep a border for high contrast.

Don't bother with the check() mixin, just inline the styles as they are
only used once and it's simpler that way. The number of added/removed lines
speaks for itself.