stylesheet: Update scales and progress bars
Use a similar style to the new checks, radios, switches and level bars.

Use an always light slider with a drop shadow, change that shadow for hover
and active. Since the shadow is not visible in dark variant, also change
the slider color in a way that will only affect dark.

Since progress bars use the same styles, update them too.