stylesheet: Update switch style
Use a similar style to the new checks and radios. Keep the slider light
even in dark variant for better visibility, but make it slightly darker
when not checked or hovered.

Unlike with checks, a border doesn't work well with HC here, so instead
use the border color as a background and make the symbols always white.

Change the trough colors on hover, since the whole switch is clickable and
not just the slider.

Add a slight drop shadow to the slider, drop it when insensitive.