Release 1.0.0-alpha.1
Use SemVer as our versioning scheme, and release libadwaita
1.0.0-alpha.1 as the first alpha version of the libadwaita 1 API. The
first beta will be 1.0.0-beta.1 and the first stable will be 1.0.0.

Given the library was unreleased and pre-alpha, this version number
change isn't regarded as a break but as a fix, and applications using
the library and checking its version number are held responsible for
updating their code.

Fixes https://gitlab.gnome.org/GNOME/libadwaita/-/issues/44