stylesheet: Use a more contrasty accent for foreground
Use lighter color in dark variant and darker color in light variant. Since
this was the main reason to lighten accent for dark, use the same color as
in light variant now, this makes it simpler and widgets like suggested
action buttons don't look too bright now.

We still need a way to easily make labels accent-colored using the correct
color though as it was a prominent use case for this. It will be added in
the next commit.