Add AdwFoldThresholdPolicy
Currently we have hardcoded behaviour for flap and leaflet,
where leaflet uses the sum of natural widths and flap uses
the sum of minimum widths to determine what size to fold
at.

Some applications, like Fractal, need to be able to adjust
this behaviour. For that purpose, we need AdwFoldThresholdPolicy.