tab-view: delay notify::selected-page until after setting
Currently, notify::selected-page can be emitted before the item is
available in the wrapped GtkSelectionModel as it has not yet been
notified. This causes NULL dereference when used with tooling such as a
GtkListView.

To fix this, we need to set the selection internally, then emit the notify
signal after the wrapped selection model has been updated.

Fixes #235