stylesheet: Overhaul header bar and tab bar borders
Use the same trick with negative margins as we've been using with solid-csd.
It's not particularly good but since nobody had complained about it for
years, should be good enough.

This should also fix the long-standing issue of discolored separators between
recolored header bars.

Use an explicitly opaque color for tab bar, we'll need this later as tab bar
cannot work with transparent borders at the moment because of the darker tab
bar background.