Implement get_request_mode() where appropriate
When using a layout manager, get_request_mode() comes for free, but
otherwise it has to be implemented manually for every widget.

Coincidentally, we've had a copy of GtkLayoutManager's get_request_mode()
in AdwHeaderBar and AdwViewStack. Extract it into widget utils and use it
everywhere.

Fixes https://gitlab.gnome.org/GNOME/libadwaita/-/issues/70