stylesheet: Add back colors like theme_selected_bg_color as aliases
We're gonna replace theme_bg_color and theme_base_color in the following
commits. Removing those from public colors would break a lot. so we should
keep those and point them to whatever colors we actually use.

And with that in mind, let's add back the ones we've already removed as
aliases.