Add AdwButtonContent
We're going to use buttons that use icon+label more commonly, so we need
a nice way to create them.

Unfortunately for technical reasons we can't reuse .icon-button.text-button
style classes, so we'll have to roll our own .image-text-button. We also
still have to add a style class on a button so that later we can know to
flatten these buttons.

Update the existing .icon-button.text-button style as well.