tab-box: Support recolorability better
Move background to a separate widget instead of tab box itself.

Make tabs clip other tabs and the new background. This means that we can
have half-transparent background and borders on both, and let the
background of the tab bar be seen through.

This doesn't adjust colors yet, that will be done later when refreshing
colors in general. However, we can already get rid of the mix()ed border.