stylesheet: Remove .sidebar border
It's been a problem for a long time, and now that designers want consistent
sidebar colors it's gonna get in the way even more as if you just ignore it
and use a plain listbox/listview as a sidebar it will have the wrong color.

So just remove that border and tell people to use manual separators
instead.

Keep an explicit border in GtkAssistant. Inspector and gtk demos will be
missing borders, but not much we can do here.

Update migration guide.