demo: Add a style classes demo page
Currently style classes are easter eggs. Some of them are documented in GTK
widgets (e.g. button.suggested-action), but most aren't, and the ones
libadwaita adds can't be documented like that. Later we'll have stylesheet
docs for this, but for now let's have at least some reference.

This is missing a few style classes, but most should be present.