view-switcher-bar: Unset stack in dispose()
Make sure we won't get an update_bar+revealed() call during dispose right
after having unparented the action bar but before it's set to null. That
only works by accident because we have a revealer that we do set to null,
but it will become a problem in the next commit.