swipe-tracker: Split begin-swipe and prepare
Move begin-swipe to when the swipe actually starts, replace what it was
before with a prepare signal.

Currently we have a hack in AdwFlap to check starting swipe in update-swipe
instead of begin-swipe because the latter runs too early. Now it can be
removed.

Update migration guides.

Fixes https://gitlab.gnome.org/GNOME/libadwaita/-/issues/79