general: Depend on glib 2.66 for now
It's not critical to depend on 2.70 right now, and it can cause problems:
even some of our CI images don't have 2.70 yet.

Instead, define G_DEFINE_FINAL_TYPE and G_DEFINE_FINAL_TYPE_WITH_CODE for
older glib versions, so we can use the newer macros when available.

This should be good enough for now, later if we decide to bump the version
this can be reverted.