action-row: Allow suffixes to expand
GtkBox packs children from the start to end. Unfortunately, here we want
to have the opposite so we need to set halign=end on the suffixes box.

However, this also breaks hexpand for its children. There doesn't seem to
be a clean way to accomplish it without changing GtkBox, so instead do a
little hack: wrap the box into another widget and that would check if the
child is expanded and control its halign.