spring-animation: Fix clamp for in-place animations
We initially went with clamping the side opposite the initial velocity.
Unfortunately, that doesn't really make sense when the user needs to just
prevent overshooting and doesn't specifically differentiate between regular
and in-place animations. So instead skip it altogether in this case.

Fix a logic error with incrementing as well.