clamp-layout: Fix measure()
So turns out after all the fixes this was still wrong:

- When get_child_size() gets for_size == -1, it returns the natural size
of the child. That's all good and well, but in reality we need the natural
size of the clamp itself - which would be larger because of the padding at
that size. So we need to calculate it - reverse the lerp and ease out cubic
we do on the clamp size to get the child size.

Refactor everything, extract inverse clamp function since we use it in two
places now.

Fixes https://gitlab.gnome.org/GNOME/libadwaita/-/issues/347