tabs: Separate clipping into another widget
When opening and closing tabs, we animate their width from/to 0. However,
tabs have padding and border, and so doing that is not allowed.

Additionally, AdwTab allocates its contents out of bounds and set
overflow=hidden.

Move all of that into gizmos that AdwTabBox would manage, and leave padding
and borders on AdwTab. That way the widget that gets animated to 0 width
and that allocates its child out of bounds has no border or padding and the
errors go away.

Fixes https://gitlab.gnome.org/GNOME/libadwaita/-/issues/352