build: Redo gen-public-types.sh in Python
This way, we make builds easier in environments that do not have a unix-like
shell and/or sed available, such as on Visual Studio and clang-cl.

For Visual Studio builds, this means that we no longer need to have a separate
Cygwin or MSYS2 installation to carry out the build.