preferences-row: Add a "title-selectable" property
For the GTK4 port of Contacts, we're using `AdwPreferencesRow`s and
`AdwActionRow`s.

In previous versions, we just used a `GtkLabel` to present the row. We
could then set the `"selectable"` property to TRUE, to deal with a major
use case when people use Contacts: to actually copy info (like a phone
number) into whatever other application they are using.

With `AdwPreferencesRow` that's not possible, so this would be a
regression.

Fixes https://gitlab.gnome.org/GNOME/libadwaita/-/issues/300