Fix AdwToast's use of GVariant
The adw_toast_set_action_target_value() function only
acquires a reference on the GVariant, but we use it
internally with variants directly returned by a
g_variant_new constructor, which returns a floating
reference that must be sunk.

Functions that take a GVariant should always call
g_variant_ref_sink(), unless they are merely holding
a reference and they are documented as requiring a
non-floating GVariant.