adw-view-switcher-title: Clear pending idle callback on unrealize
It’s possible (but quite unlikely) for the widget to be unrealized and
destroyed while `check_window_width()` is pending in an idle callback.

If that happens, `check_window_width()` will be called on a disposed
`AdwViewSwitcherTitle` instance, and will crash.

Avoid that by explicitly removing the idle source when the widget is
unrealized.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>