entry-row: Hide apply button when disabled
When setting the :show-apply-button property to FALSE, it seems
reasonable to expect that the button will not be shown, regardless
of whether it was visible before or not.

This is done by resetting the text_changed state. As that is only
tracked if the property is set, resetting it makes sure that
toggling the property behaves consistently.