tab-view: Use global scope for shortcuts
Per discussion in #gtk - we want these shortcuts to be handled by the
window. In practice, `MANAGED` usually does the same as `GLOBAL`: it
would normally use a GtkWindow since that's the common
GtkShortcutManager ancestor. `MANAGED` is mainly useful for menus,
so it's more appropriate for us to use `GLOBAL`.