tab-view: Add shortcut controller with capture phase
Certain apps and widgets interfere with our PgUp and PgDown
shortcuts. Particularly, these shortcuts don't work in Epiphany
or Nautilus because WebKit and GTK's new View widgets block
or use them. To fix this, we need to use the capture propagation
phase. However, we don't want to use capture for every single
shortcut.

This commit adds a separate controller to be used for the PgUp
and PgDown shortcuts to resolve the issue.