entry-row: Don't select text when clicking the row
While the internal GtkText spans most of the row, it doesn't cover all of
it. In particular, this is noticeable with prefix or suffix widgets.

When clicking inside the GtkText, it just focused it and moves the text
cursor. However, when clicking outside it, it also selects the text. Change
that so it feels less random.