dist-data: Error out if it can't find docs
The purpose of that script is to include pre-built docs and stylesheets
into the tarball. However, while building styles is automated and the
script does it on its own, the docs need to be manually built beforehand.
This step is easy to miss, so visibly fail instead of continuing with an
incomplete tarball.