src/meson.build: Look for Vulkan as well
Vulkan support may have been enabled for GTK but is only pulled in with the
GTK dependency when it is found via pkg-config, which will break the build
if Vulkan was found via other methods, which is typical on Windows as the
LunarG SDK sets envars instead of installing pkg-config files.

This will ensure that libadwaita continues to build properly in this case
so that the Vulkan headers can be found during the build, which are pulled in
by the GTK headers when it is built with Vulkan support.