preferences-window: plug leak in make_comparable()
`pango_parse_markup()` is used to overwrite a pointer to an allocated
string. Since the input to this function is transfer-none and the
output is transfer-full, this leads to a leaked string.

See also GNOME/gi-docgen#141 for why this probably happened.


(cherry picked from commit 157aecac7b04b04b989da7bd763911ccfd498452)