message-dialog: Stop using AdwSqueezer
When originally implementing the dialog, I took a shortcut and used
squeezer instead of reimplementing the layout manually, and bit the
bullet with disappearing focus when switching layouts.

Since squeezer is to be deprecated, rework that and use a proper custom
layout.

This also allows to keep focus on the correct button after switching.