version: Update deprecation macros
We need to combine them with _ADW_EXTERN to handle exporting. Along the
way, align them with GDK, adding per-version deprecations, and drop the
duplicate macros we've never used. Drop hdy-deprecation-macros.h and merge
the 2 remaining macros into adw-version.h as we need to declare them after

This was done in libhandy 1.5 years ago, and somehow never made it to
libadwaita.