tab: Add extra-drag-value signal
The extra drop target does not allow for dynamic changing of the
preferred action.  The preferred action is set on a GtkDropTarget
from the return result of ::enter and ::motion.  Without this ability,
a extra drop target with multiple allowed actions is not very useful.
extra-drag-value allows the possibilities to see your source value
as well as the destination (AdwTabPage) in order to determine the
preferred action which will get passed to GtkDropTarget::enter/motion.

This really should be called extra-drop-value, but opted for
extra-drag-value to keep consistent with ::extra-drag-drop.