tab: Notify value on ::enter
The ::extra-drag-value signal handles local DnD very well
as the :value property becomes available almost immediately.
For non-local drags, however, there may be a noticeable
amount of time between when the drag enters the
drop target, and when the value is ready.  During this
time, the application doesn't have any control over
the preferred action.

Emit ::extra-drag-value on ::enter, so that the application
has control over the action (particularly when the value
is NULL).  This has the added benefit of making the
application more aware when preload is mistakenly set to FALSE.