toast-widget: Don't focus buttons on click
See https://gitlab.gnome.org/jrb/crosswords/-/issues/137

This does not completely fix that issue in general, but it does prevent it
from happening in crosswords specifically, and there's no reason to focus
toast buttons on click anyway - it only causes focus to get lost once the
toast is gone.

If this pops up in some other app again we can look at it again, for now
this is good enough.