carousel: Don't skip allocation if not realized
No idea why this check was here or what the implications are; Alex says
this might be a GTK 3 leftover. Either way, in tests, the widgets are
not realized, and this early return causes Carousel's child widgets to
be allocated with a size of 0×0 instead of the correct size, causing
failures.


(cherry picked from commit cbfd8e57d5ba684c8744f6955e0435e7a09b993e)