settings: Be less greedy with color-scheme handling
When originally implementing this, the idea was to be as robust as possible
and support as many ways to fetch the preference as we can. In practice,
this ranges from being useless to outright harmful, as it causes
:system-supports-color-schemes to be set to true when it shouldn't.

We still need GSettings handling, but only for xdg-desktop-portal-gnome, so
restrict it to ADW_DISABLE_PORTAL=1.

This will probably make some people without portals on their system unhappy
that libadwaita apps stopped picking up gsettings, but they can pass
ADW_DISABLE_PORTAL=1. Since they don't have portals anyway, they aren't
losing anything.