preferences-page: Add "description" property
Adds a new "description" property, a corresponding test, and a demo to the
AdwPreferencesPage. It is shown as a description text at the top of the
page, if present. This pattern is common in GNOME Settings and is
usually implemented by having the first element of AdwPrefenecesPage to
be an empty AdwPreferencesGroup with only a "description" property set.
This change also makes the description text have proper margins as per
mockup.