adw-tab-bar: Add property extra-drag-preferred-action
setup_extra_drop_target() allows for multiple actions on the
extra drop target.  The preferred action can be changed from the
result of ::extra-drag-value and using the keyboard modifiers
CTRL/SHIFT.  We never have a way of seeing what the preferred
action is because we don't get the GtkDropTarget in ::extra-drag-drop.
Add a property extra-drag-preferred-action for the purpose of
getting the preferred action in ::extra-drag-drop.

Fixes: https://gitlab.gnome.org/GNOME/libadwaita/-/issues/660