view-switcher-bar: Disable reveal animation when breakpoints are used
That animation can still play in a lot of cases, and it really shouldn't.

In 2.0 we should just remove the reveal property entirely and tell people
to use :visible instead. If they want an animation for other purposes, they
can use a revealer instead. For now we're stuck with that property, but can
at least disable animation.