back-button: Add accessible label
This got lost when we switched to ::query-tooltip.

Only have "Back" as a label instead of the name of the destination view, as
it would be a nightmare to keep that up to date at all times, and "Back" is
probably a better choice for accessibility anyway.

Fixes https://gitlab.gnome.org/GNOME/libadwaita/-/issues/715