expander-row: improve the accessible role
Set the accessible role of the header row's `GtkListBox` container to
"generic" inside of "list". This results in a screen reader reading
"Title, not selected, collapsed, Subtitle" instead of reporting a
"List with one item".

The latter is confusing, since the expander row itself will almost
always be in a list whose length won't be announced.