combo-row: Preserve focus when opening the popover
When the tab-behavior is set to all, after a widget is shown, a tab key
press is emulated to focus the list view. Because of that, the first item
always gets focus.

Fix it by setting the property to item, which fixes it, and it makes more
sense from the point of keyboard navigation as well.