tab-overview: allow child focus on out animations
Previously, AdwTabOverview could not focus the child until the overview had
completely animated out. This causes problems for applications because they
may end up in a state at the end where no widget is grabbed.

Additionally, to work around this with timeouts to forcefully refocus a
child after the animation completes will lose input in the mean time. For
applications that are input heavy (editors, terminals, etc) that can be
quite frustrating from a keyboard perspective.

This seeks to avoid that by treating animation in and out differently. This
allows focusing the child so long as we're on the way out with the
animation and therefore input is delivered as expected.

Related #670


(cherry picked from commit 42c04e038f19b2123560da662692d65480a67931)

(cherry picked from commit 75f49e900aa6a23509cc5db48db589d03184fe58)