alert-dialog: Fix initial focus
Previously it was using an adapted version of AdwMessageDialog code, which
in turn was copied from GtkDialog. I have no clue what that code does aside
from the vague "sets focus", and it didn't work properly anyway.

Throw all of that away and rewrite it from scratch. Now it actually works,
and is readable too.