dialog-host: Use ancestor for already-unmapped focus widget
Specifically, this can happen when opening a dialog from a menu. The focus
widget was a menu item, and then menu closes and it's unmapped, so we've
lost the focus entirely. Go up in the hierarchy and remember the menu
button instead.

Fixes https://gitlab.gnome.org/GNOME/libadwaita/-/issues/830