stylesheet: support GtkGridView and GtkFlowBox as a .navigation-sidebar
Just like lists, grids can be used to navigate in an app.
An example for this would be the page overview in Papers'
sidebar.

And, while to the best of my knowledge, Papers is
currently the only user grids there and thus benefitting
from this, conceptually, this is not special to it.

Closes: https://gitlab.gnome.org/GNOME/Incubator/papers/-/issues/36

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>