back-button: Update on AdwNavigationPage::showing instead of ::pushed
When showing the page, we try to focus a widget in it before the back
button is visible. If there are no other focusable widgets, we fail, and
that's not good.

Make sure to show the (initially hidden) back button when showing the page,
before focusing anything.

Meanwhile we don't need the pushed signal anymore. That's what it was
originally for, but it's emitted too late to be useful here.


(cherry picked from commit 465bd102c1eb61b2e84e9bb4762701b558d558df)