docs: Use getter and setter gir annotations
These are needed only when the name of the getter or setter for a
property called "AdwClassName:property-name" is not called
"adw_class_name_{g,s}et_property_name".

The getters (setters) were found by regenerating the rust bindings with the following
check:

if org.gtk.Property.g(s)et is set but g(s)etter isn't, then print a message.

and the same for get-property and set-property.