combo-row: Don't change row sensitivity on model change
AdwComboRow becomes sensitive or insensitive automatically when the
model changes. In particular, it is forced sensitive when the model
contains more than 0 items.

This can mess with the user-set sensitivity of the row if the model
changes. See gnome-control-center#3122.

So let's remove that sensitivity management althogether. The row is
anyways not activatable once it's empty and so can't be interacted with
furter.