spinner: add workaround for memory corruption issue
This is a workaround for
https://gitlab.gnome.org/GNOME/glib/-/issues/3434 which is a duplicate
of https://gitlab.gnome.org/GNOME/glib/-/issues/1002.

The problem is GObject atomically destroys its list of weak notifies
when an object is disposed, *then* iterates through each notify. If one
notify tries to remove a later notify, it will fail and emit a critical.
Then the later notify will be improperly executed, leading to use after
free.

There is a solution in
https://gitlab.gnome.org/GNOME/glib/-/merge_requests/3864, which changes
GObject to remove weak notifies just one at a time. But that is going to
take a while to land, and it's a longstanding bug. We should land a
workaround now so that applications can safely use AdwSpinner in the
meantime.