test-bottom-sheet: Pass alignment in varargs correctly
The type of the property is declared to be G_TYPE_FLOAT, therefore it
must be passed on the stack as a float (which in practice is promoted
to double via the "default argument promotions" as defined by Standard C).

Previously the alignment was written to the stack as a 4-byte integer,
but read off as an 8-byte double, causing undefined behaviour (in
practice dependent on uninitialized stack contents).

Resolves: https://gitlab.gnome.org/GNOME/libadwaita/-/issues/913
Signed-off-by: Simon McVittie <smcv@debian.org>