preferences-page: Add banner property
Add a banner property to AdwPreferencesPage. This AdwBanner widget is
shown at the top of the page if it is revealed and does not scroll with
the page contents. The intended use is to draw attention to conditions
that might affect the preferences on the page, such as downloads being
paused on a metered network.