bin: Delay child's parent check
AdwBin checks if the child has a parent before checking if it is
its own child. That means subsequent calls to adw_bin_set_child()
with the same widget will trigger a warning unnecessarily.

Delay the child's parent check to after checking if the child is
the bin's own child.