clamp: Stop overriding get_request_mode
Let the default implementation in GtkLayoutManager take over.
Specifically, it will propagate child's preferred request mode.

Generally, it's best to avoid measuring a widget against its preferred
size request mode, for two reasons:
1. It may be extremely slow, since it will commonly have to do (multiple
   layers of) binary searching.
2. Many widgets have a not-quite-strictly-correct measure implementation
   for the opposite size request mode. Indeed, oftentimes the only way
   to implement correct measurement is to do a binary search.
As a relevant example of the latter point, AdwWrapLayout itself has an
incorrect implementation of measuring in the opposite orientation, which
will be fixed over the next few commits.

This is why it is generally a very good idea to propagate the child's
request mode, unless there are good reasons why a container widget
strongly prefers to be measured in a specific orientation. GtkListBox
is a good example of a case where it's reasonable to prefer height-for-
width (which is likely to match the children's preferred mode anyway),
while the regular GtkBox (intended to be used with a small number of
child widgets) should propagate its children's preferred request mode,
even though it's generally easier for a box itself to be measured along
its own orientation, not across it.

AdwWrapLayout can be quickly and nicely measured in either orientation,
so don't force a request mode and propagate what the child asks for. As
a bonus point, we will report constant-size when the child is, which
enables further optimizations elsewhere in GTK.

Signed-off-by: Sergey Bugaev <bugaevc@gmail.com>