clamp: Round child size down, not up
If both child_size_from_clamp () and clamp_size_from_child () round up,
the measurements are inconsistent between the two directions. It can be
shown mathematically that the proper arrangement is to round down when
converting from a container size to its child's size, and round up when
converting from the child's size to the container size.

So make child_size_from_clamp () use floor () in stead of ceil ().

Signed-off-by: Sergey Bugaev <bugaevc@gmail.com>