clamp: Fix measuring along the clamped orientation
The existing code attempted to base the clamping calculations of the
minimum and natural sizes of the child for the given size in the
opposite orientation. While it's generally a good idea to pass the
appropriate for_size value instead of -1 to gtk_widget_measure () calls
whenever possible, in this case this only leads to inconsistencies with
measurements along clamp's own orientation (i.e. height-for-width for a
horizontal clamp) and with allocate () behavior.

So stop passing minimum and natural sizes to clamp_size_from_child (),
and only pass a single size to be mapped. In clamp_size_from_child (),
query the child for its overall minimum and natural sizes, and base our
calculations on those. This is consistent with child_size_from_clamp ().

Moreover, the existing code only mapped the child's natural size,
passing through the minimum size as is. This is the correct thing to do
for overall minimum size, but not when measuring for a specific size in
the other orientation: when the minimum size for the specific size in
the opposite orientation is larger than the overall minimum size and
larger than our tightening threshold, we would give the child less space
if the clamp got allocated this size, and that would contradict it being
the minimum acceptable size for the child for the given size in the
other orientation. So we must map the minimum size reported by the child
as well.

Also, fix baseline handling for vertical clamps.

Signed-off-by: Sergey Bugaev <bugaevc@gmail.com>