clamp: Fix measuring opposite minimum size
When passed for_size = -1, the existing code measured the child for
MIN (nat, ceil (max)), where nat is the natural size of the child in the
opposite orientation, and ceil (max) is the maximum size that the clamp
will possibly allocate to the child. This generally makes sense.

However, it's not always true that the overall minimum size of a widget
equals it minimum size for its natural size in the opposite orientation.
A specific example of this is GtkLabel with max-width-chars set. The
overall minimum height for such a label will be very small (perhaps the
height of a single line of text), while the minimum height for its
natural width will be considerably larger. Of course, the same applies
to complex widget trees that contain such labels.

What we really need to report here is the smallest possible size that
the child will ever have. We cannot just propagate the child's overall
minimum size, since that might be only achievable at a size in the
opposite orientation that's above the maximum size we're going to give
it. So instead use the minimum size for that maximum size in the other
orientation; this is going to be at least as small as minimum size for
any other size, since measurements are assumed to be monotonic.

Signed-off-by: Sergey Bugaev <bugaevc@gmail.com>