wrap-layout: Handle n_visible_children <= 1 specially
With no visible chidren, just return 0. This lets us report constant-
size request mode; but it's not only an optimization, it also fixes
several bugs, as some code paths failed to account for the case of
having no children.

With a single child, don't do any wrapping, and pass through the child's
measurements and the request mode as-is; if the child is constant-size,
so are we.

The specific use case I have in mind for these optimizations is showing
a list of reactions under a post or a message. Most posts would have
no reactions or a single reaction, and the widget representing a
reaction (perhaps an hbox with an icon and a label) would likely itself
be constant-size.

Signed-off-by: Sergey Bugaev <bugaevc@gmail.com>