wrap-layout: Fix measuring opposite size
compute_line was early-exiting given for_size == -1, so box_allocate was
never being invoked, and the children were not allocated any sizes. This
then caused compute_sizes to attempt to measure the children for size of
zero.

The appropriate thing to do, given for_size == -1, is to measure the
children for -1 too, so do that instead. Also nano-optimize
count_line_children for this case: instead of trying to fit widgets into
unlimited space, just say that all the widgets fit.

Signed-off-by: Sergey Bugaev <bugaevc@gmail.com>